+++ MordoEngine/res/shaders/triangle.fs
@@ -1,8 +1,11 @@
 #version 330 core
 
+in vec2 TexCoord;
 in vec3 vColor;
+
 out vec4 FragColor;
+uniform sampler2D texture1;
 
 void main() {
-    FragColor = vec4(vColor, 1.0);
+    FragColor = texture(texture1, TexCoord);
 }