+++ MordoEngine/src/Core/Texture/Texture.h
@@ -1,0 +1,10 @@
+#pragma once
+#include <glad/glad.h>
+#include <string>
+
+class Texture {
+public:
+	GLuint ID;
+	Texture(const std::string& texturePath);
+	void Use() const;
+};