+++ MordoEngine/src/Renderer/TriangleRenderer.h
@@ -2,17 +2,19 @@
 #include <glad/glad.h>
 #include "../Core/Shader/Shader.h"
 #include "Triangle.h"
+#include "../Core/Texture/Texture.h"
 #include <vector>
+#include <random>
 
 class TriangleRenderer {
 private:
-	GLuint vao, vbo, instanceVBO;
+	GLuint vao, vbo;
 	std::vector<Triangle> triangles;
 	Shader shader;
-	bool flag;
 	void AddRandomTriangle();
 	void HandleInputs();
-
+	TextureID defaultTextureID;
+	std::mt19937 gen;
 public:
 	TriangleRenderer();
 	~TriangleRenderer();