+++ docs/changelog.md
@@ -11,3 +11,12 @@
 - Added GLM for mathematical operations
 - Integrated TriangleRenderer into the engine loop for testing the input system and shader class
 - Added dynamic triangle creation using the space key in TriangleRenderer
+
+## 2026-01-27
+- Added a filesystem utility for loading shader source code from files
+- Added stb_image for texture loading
+- Implemented a Texture class for loading and binding textures using stb_image
+- Added a texture manager for efficient texture loading and caching
+- Updated TriangleRenderer to apply textures to triangles
+- Updated triangle render implementation
+- Updated shader class to set uniform variables